rename libraries, add comments
