+++ sbnobj/Common/Reco/CMakeLists.txt
@@ -3,6 +3,7 @@ cet_make_library(
     CNNScore.cc
     CRUMBSResult.cc
     FlashTriggerPrimitive.cc
+    LightCalo.cc
     MVAPID.cc
     MergedTrackInfo.cc
     OpT0FinderResult.cc
