+++ sbnobj/Common/Reco/LightCalo.cc
@@ -1,0 +1,1 @@
+#include "sbnobj/Common/Reco/LightCalo.h"
