--- sbnobj/Common/Reco/LightCaloInfo.cc
@@ -1,1 +1,0 @@
-#include "sbnobj/Common/Reco/LightCaloInfo.h"
